fix: avoid time_t overflow in timediff_in_ms()
Cast each tv_sec operand to double before subtracting to ensure
the arithmetic is performed in floating-point. The previous
(double)(later->tv_sec - earlier->tv_sec) performs the subtraction
in time_t first, which is signed overflow UB when the difference
is large or when time_t is narrow.

+++ src/time_util.c
@@ -134,6 +134,6 @@ int sleep_timespec(const struct timespec *duration) {
  */
 double timediff_in_ms(const struct timespec *later,
                       const struct timespec *earlier) {
-    return (double)(later->tv_sec - earlier->tv_sec) * 1e3 +
+    return ((double)later->tv_sec - (double)earlier->tv_sec) * 1e3 +
            ((double)later->tv_nsec - (double)earlier->tv_nsec) / 1e6;
 }
